+++ apps/laboratory/tests/smart-account.spec.ts
@@ -1,7 +1,6 @@
 import { testModalSmartAccount } from './shared/fixtures/w3m-smart-account-fixture'
 import type { ModalWalletPage } from './shared/pages/ModalWalletPage'
 import { EOA, SMART_ACCOUNT } from './shared/validators/ModalWalletValidator'
-import { Email, NOT_ENABLED_DOMAIN } from './shared/utils/email'
 
 import type { ModalWalletValidator } from './shared/validators/ModalWalletValidator'
 
@@ -65,37 +64,6 @@ testModalSmartAccount(
     await walletModalPage.closeModal()
 
     await walletModalValidator.expectAddress(originalAddress)
-  }
-)
-
-testModalSmartAccount(
-  'it should use an eoa and not propose flow when disconnecting and connecting to a not enabled address',
-  async ({ modalPage, modalValidator, context }, { parallelIndex }) => {
-    const walletModalPage = modalPage as ModalWalletPage
-    const walletModalValidator = modalValidator as ModalWalletValidator
-
-    const email = new Email(mailsacApiKey)
-
-    await walletModalPage.openAccount()
-    await walletModalPage.openSettings()
-    await walletModalPage.togglePreferredAccountType()
-    await walletModalValidator.expectChangePreferredAccountToShow(EOA)
-    await walletModalPage.disconnect()
-    await walletModalPage.page.waitForTimeout(1000)
-
-    await walletModalPage.emailFlow(
-      email.getEmailAddressToUse(parallelIndex, NOT_ENABLED_DOMAIN),
-      context,
-      mailsacApiKey
-    )
-    await walletModalPage.page.waitForTimeout(1500)
-    await walletModalPage.openAccount()
-    await walletModalPage.openSettings()
-    await walletModalValidator.expectTogglePreferredTypeVisible(false)
-    await walletModalPage.closeModal()
-
-    await walletModalPage.openAccount()
-    await walletModalValidator.expectActivateSmartAccountPromoVisible(false)
   }
 )