+++ packages/scaffold-utils/src/ConstantsUtil.ts
@@ -14,5 +14,5 @@ export const ConstantsUtil = {
   CONNECTOR_RDNS_MAP: {
     coinbaseWallet: 'com.coinbase.wallet'
   } as Record<string, string>,
-  VERSION: '4.2.3'
+  VERSION: '5.0.2'
 }