+++ packages/scaffold/src/partials/w3m-social-login-widget/index.ts
@@ -2,6 +2,7 @@ import {
   AccountController,
   ConnectorController,
   CoreHelperUtil,
+  EventsController,
   RouterController,
   SnackController,
   type SocialProvider
@@ -23,9 +24,7 @@ export class W3mSocialLoginWidget extends LitElement {
   // -- Members ------------------------------------------- //
   private unsubscribe: (() => void)[] = []
 
-  private desktopWindow?: Window | null
-
-  private mobileWindow?: Window | null
+  private popupWindow?: Window | null
 
   // -- State & Properties -------------------------------- //
   @state() private connectors = ConnectorController.state.connectors
@@ -153,16 +152,19 @@ export class W3mSocialLoginWidget extends LitElement {
   async onSocialClick(socialProvider?: SocialProvider) {
     if (socialProvider) {
       AccountController.setSocialProvider(socialProvider)
+      EventsController.sendEvent({
+        type: 'track',
+        event: 'SOCIAL_LOGIN_STARTED',
+        properties: { provider: socialProvider }
+      })
       RouterController.push('ConnectingSocial')
     }
     const authConnector = ConnectorController.getAuthConnector()
-    if (CoreHelperUtil.isMobile()) {
-      this.mobileWindow = CoreHelperUtil.returnOpenHref(
-        '',
-        'popupWindow',
-        'width=600,height=800,scrollbars=yes'
-      )
-    }
+    this.popupWindow = CoreHelperUtil.returnOpenHref(
+      '',
+      'popupWindow',
+      'width=600,height=800,scrollbars=yes'
+    )
 
     try {
       if (authConnector && socialProvider) {
@@ -170,23 +172,16 @@ export class W3mSocialLoginWidget extends LitElement {
           provider: socialProvider
         })
 
-        if (!CoreHelperUtil.isMobile()) {
-          this.desktopWindow = CoreHelperUtil.returnOpenHref(
-            uri,
-            'popupWindow',
-            'width=600,height=800,scrollbars=yes'
-          )
-        }
-        if (this.desktopWindow && uri) {
-          AccountController.setSocialWindow(this.desktopWindow)
-        } else if (this.mobileWindow && uri) {
-          this.mobileWindow.location.href = uri
-          AccountController.setSocialWindow(this.mobileWindow)
+        if (this.popupWindow && uri) {
+          AccountController.setSocialWindow(this.popupWindow)
+          this.popupWindow.location.href = uri
         } else {
+          this.popupWindow?.close()
           throw new Error('Something went wrong')
         }
       }
     } catch (error) {
+      this.popupWindow?.close()
       SnackController.showError('Something went wrong')
     }
   }