+++ packages/scaffold/src/views/w3m-connecting-social-view/index.ts
@@ -3,6 +3,7 @@ import {
   AccountController,
   ConnectionController,
   ConnectorController,
+  EventsController,
   ModalController,
   RouterController,
   SnackController,
@@ -128,21 +129,48 @@ export class W3mConnectingSocialView extends LitElement {
 
             if (this.socialProvider) {
               StorageUtil.setConnectedSocialProvider(this.socialProvider)
+              await ConnectionController.connectExternal(this.authConnector)
+              EventsController.sendEvent({
+                type: 'track',
+                event: 'SOCIAL_LOGIN_SUCCESS',
+                properties: { provider: this.socialProvider }
+              })
             }
-            await ConnectionController.connectExternal(this.authConnector)
           }
         } catch (error) {
           this.error = true
           this.updateMessage()
+          if (this.socialProvider) {
+            EventsController.sendEvent({
+              type: 'track',
+              event: 'SOCIAL_LOGIN_ERROR',
+              properties: { provider: this.socialProvider }
+            })
+          }
         }
       } else {
         RouterController.goBack()
         SnackController.showError('Untrusted Origin')
+        if (this.socialProvider) {
+          EventsController.sendEvent({
+            type: 'track',
+            event: 'SOCIAL_LOGIN_ERROR',
+            properties: { provider: this.socialProvider }
+          })
+        }
       }
     }
   }
 
   private connectSocial() {
+    const interval = setInterval(() => {
+      if (this.socialWindow?.closed) {
+        if (!this.connecting && RouterController.state.view === 'ConnectingSocial') {
+          RouterController.goBack()
+        }
+        clearInterval(interval)
+      }
+    }, 1000)
     window.addEventListener('message', this.handleSocialConnection, false)
   }